Synthesised audio and compared sample rates
Sounds of individual bubbles are generated from their second order
differential equations. Each sound is time shifted and added to create
the sound of many bubbles. I use a matix of source data to generate the
sounds, the advantage of this is that I can use that source data to
compare my bubble detection algorithm to the true values of radius and
time. Synthesising the audio is slow, because it involves solving second
order differential equations, so to improve performance, audio data is
memoized (cached), so that when I rerun the script, it can skip these
calculations.

I compared sample rates of 22050, 44100 and 88200Hz, and I did not see a
noticable effect on the quality of bubble detection, suggesting that the
only effect is on nyquist frequency or minimum bubble radius detectable.