Try custom wavelets in MATLAB
This attempts to add a custom wavelet to MATLAB to use for creating cwt
scaleograms. The wavelet added is a SOULTI wavelet - one based on a
second order LTI system. I chose to try this as the bubbles are a second
order LTI system so I am speculating that this type of wavelet may be
more effective than the generic inbuilt one.

Unfortunately, it appears that the MATLAB does not support user defined
wavelets. There documentation online that suggests it does is for an
older version of the cwt function, while the current version does not.