Implement source separation for three hydrophone simulation.
Audio from the three hydrophones is combined using a simple source
separation algorithm (delay and add). The aim of this is to increase the
sound intensity of audio originating from bubbles at the specified
location, and attenuating the audio of bubbles outside that area.

Unfortunately, because the speed of sound in water is so fast, the delay
and add source separation algorithm is not very good at attenuating
audio from outside the desired area. This is because the delay time
between hydrophones is very short, much shorter than the period of the
signal.

As a result, multi hydrophone setups are of no advantage, unless they
were spaced much further apart.