Add simple script example
This adds a script called bubble audio, which performs bubble
identification on a single hydrophone at a single sample rate. This is
not particularly useful as a simulation as it does not compare anything
or teach me anything, but it is a good example for explaining how the
program works. Useful as a reference for creating more complicated
simulations.