Implement SOULTI wavelet
This adds the SOULTI wavelet, which is a wavelet based on the response
of a second order LTI system. I wanted to see if this would be effective
at identifying the bubble sounds, since bubbles oscillate according to a
second order differential equation as well. The paper which I used as a
reference specifies two alternative SOULTI wavelets. They differ by the
value of a constant P. I use a P value of 1/2, because it keeps the
power the same across all frequency bands, which makes better looking
scaleograms. Using P = 1 is required for reconstructing a signal using
the inverse transformation, but for my application, I don't care about
that.