Fix time reversed wavelets and distortion of cwt
This corrects an issue where wavelets were time reversed. The SOULTI
wavelet is not time symetric, and so this issue affects the resulting
scaleograms. Using the correct SOULTI wavelet results in a cleaner and
sharper plot.

CWT plots were also time shifted by the length of the wavelet. This
resulted in a distortion that was more noticable at lower frequencies.