Create rust implementation of CWT
MATLAB code cannot be run on a microcontroller and is difficult to
interface with other software. Therefore there is a need to write an
implementation of my bubble identification code in another language.
This commit begins the work on creating a rust implementation of a CWT
function. To have data to feed my program, I have decided to export the
synthesised audio from MATLAB as a wav file.