Improve rust implementation of CWT
This improves the CWT functionality of my program. The iterator for
creating ranges of floating point numbers is moved to a separate file
and unit tests have been added to verify its correctness. The output is
now written out to a csv file. Numbers are in scientific notation (e.g.
1.3e4). Convolution is implemented with three different methods, one in
frequency domain, using the fast fourier transform algorithm and two in
time domain, one which uses a parallel iterator to perform multiple
operations simultaneously. The quickest function for convolution is the
one that uses parallelism, while the FFT and regular basic versions are
slower. FFT is not much faster than I expected, which might be because
the wavelet is significantly shorter than the signal length. Convolution
in frequency domain is comparatively quicker than in time domain when
the two signals are of similar length.