Partially implement convolution using SIMD
SIMD (single instruction multiple data) is a group of CPU instructions
that perform an operation on multiple values simultaneously. I plan to
use SIMD to speed up convolution, which is very slow. I started with the
faster library, but I had difficulties, so I switched to the packed simd
library. This feature is still a work in progress.