Try convolution using SIMD with i16 data types
If I use 16 bit int data types to represent numerical data instead of
32-bit floats, I can perform operations on twice the amount of data per
SIMD instruction. I would expect that this will double the speed of the
CWT function.

When I tested this, it did in fact have about double the speed. On a
release build on my computer, the CWT function was able to process 100ms
of audio in 88ms. This is a great result as it proves that real-time
bubble analysis using the continuous wavelet transform is possible. I
expect by reducing the number of reallocations of memory by reusing
arrays, I shall be able to improve performance a little further. This
gives me a budget of at least 10ms to perform other operations such as
peak finding.

However, the issue with using a fixed point data type, is that the
working range is much less. Looking at the plot, it is clear that there
are overflows. I need to limit the range of my signal and wavelet
to avoid overflows. I have positive results when I limit the wavelet to
2^2 and the signal to 2^8.