Add complex wavelets
As I found in my rust program, complex wavelets are good because they
reduce the horizontal banding in the plots. When using a real wavelet
based off of a sine wave, on a signal that is also a sine wave, when the
wavelet and the signal are in phase, they have a larger correlation and
this produces a peak on the plot. When they are out of phase, they have
low correlation and this produces a trough. This creates the banding.
Complex wavelets can be thought of as two wavelets, 90° out of phase.
Whenever one produces a peak, the other produces a trough, and the
banding produced from each cancel out. By taking the absolute value of
the wavelet plot, I can produce a plot with minimal banding, which is
very useful as banding makes peak finding more difficult.

There is still slight banding on the plots, I will still need to modify
the peak finding algorithm to ignore these.