Refactor cwt module to use an object oriented programming
This refactors the various cwt functions into objects. The advantage of
this is that this allows me to remember data, and save on processing
time. The filterbank algorithm performs the fourier transform on the
wavelet function once and stores this data in a filterbank. This speeds
up the processing time, because the transformation does not have to be
re-calculated each time new data comes in. When testing on a signal of
duration 1000 ms, with frequency bands from 1000 to 9000 Hz (20 Hz
interval), processing time was 1971 ms as compared to 2714 ms before.
This is a speed up of about 30%. However, it is not yet real-time, the
target for that is less than 1000 ms. I also need to see how the above
scales with different length signals. Does a signal of two seconds take
twice as long to process? This will determine the optimal size of chunks
to process.