Implement an iterator that takes from a circular buffer
In the future, there will be a need to have a circular buffer between
the data acquisition code and the cwt analysis code. To reduce data
copying, I plan on passing a lazy iterator into the process/process_par
method rather than a vector copy of the data. This commit implements an
iterator for a circular buffer. In a segmented cwt, there is some
overlap between reading of chunks of data. The overlap is the length of
the wavelet minus one, at the largest scale (lowest frequency band). The
circular buffer needs to advance by the length of the chunk minus the
overlap. Therefore the iterator lets you specify the amount of elements
to take and the amound of elements to peek. Unit tests for this iterator
are passed.