Generate 10 seconds of audio data using the export_to_wav m file
Soon I will be testing a segmented CWT, and to do so, I will need to
stream in a longer audio file than the one I am currently using. This
changes the export_to_wav m file to generate 10 seconds of audio data.
To support this change, the generate_audio.one_hydrophone function had
to be altered to accept a duration parameter, rather than use a constant
duration of 1.5 seconds. The bug in the make_sound function, where NaNs
or infinities would sometimes occur has also been fixed. The issue was
when a large negative time was multiplied with a negative number, and
was put into the exponential function, this would create infinities.
This is now fixed.