Implement a segmented CWT using overlap-save
Overlap-save (AKA overlap-discard) is a technique used in real-time
signal processing, for algorithms that work on segments and which need
to look ahead. A chunk is processed at a time by the CWT algorithm, and
the portion at the end of the chunk (which overlaps with the next) is
discarded. This is because continuous wavelet transforms have
distortions near the edge, where the wavelet looks ahead to future
signals.

Overlap discard is currently implemented for the FftCpxFilterBank
algorithm only. More work is needed, to implement this in others. A
'bounded' convolution module will need to be created to replace the
current conv module.

Previously, I suspected that the segmenting of the signal was not
working correctly. I have tested it further, and I now think that it is
working fine.

Benchmarking code will need to be rewritten to be compatible with the
new changes.