Compare the effects of median filtered data to non-filtered data.
Median filters are good at smoothing out data. The
compare_median_filtering m file tests out the effects of median
filtering the signal, and median filtering the 2D scaleogram. This did
not help with bubble identifications, except a median filter of length 3
on the signal, which may have had a slight positive effect, but more
testing is needed.