Plot source data and results as a scatter plot for comparison
The matlab m file scatter_plot_from_csv.m plots multiple segments of csv
data on the same plot. To assist in readability, adjacent segments
alternate between four different colours, so that it is easy to
distinguish different segments. The find_peaks function in the program
has been modified, so that the output values are radii in mm and times
in ms.