Apply mean shift clustering on algorithm data
Mean shift clustering is a clustering algorithm which does not have a
fixed or specified number of clusters - unlike K-means clustering. This
is good for this data set because we don't know how many bubbles are
expected in each segment. This reduces the number of bubble
identifications, when there is a lot of identifications very close
together.

The article which helped me choose the clustering algorithm was
https://towardsdatascience.com/the-5-clustering-algorithms-data-scientists-need-to-know-a36d136ef68.
I used the psuedocode from this paper: https://github.com/LorenzoAgnolucci/MeanShiftClustering/blob/master/paper_mean_shift_PC.pdf
as a reference for my implementation.